scripts/requirements.txt
mkdocs
mkdocs-material
pyyaml
pytz
stashy
threadly
